[libunwind] Implement a baseline sandbox using otypes.
This commit adds a new compile-time option (and enables it by default)
called LIBUNWIND_SANDBOX_HARDENED. This causes libunwind to seal all of
the capabilities it finds on the stack before returning control to the
caller. Furthermore, it ensures that a compartmentalizing RTLD returns a
sealed pointer to the trusted frame, rather than an unsealed one that
leaks to the caller.

This sandboxing model is experimental and should not yet be taken to be
secure, as there are a number of known deficiencies with it.